Add implementation of service
